+++ docs/whats-new.rst
@@ -5,6 +5,8 @@ What's new
 0.4 (*unreleased*)
 ------------------
 
+- Added the :py:func:`pint_xarray.expects` decorator (:pull:`143`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_ and `Justus Magin <https://github.com/keewis>`_.
 
 0.3 (27 Jul 2022)
 -----------------